+++ internal/api/handlers/management/oauth_settings.go
@@ -1,0 +1,149 @@
+package management
+
+import (
+	"encoding/json"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+	"github.com/router-for-me/CLIProxyAPI/v6/internal/config"
+	oauthsettings "github.com/router-for-me/CLIProxyAPI/v6/internal/management/settings/oauth"
+	"github.com/router-for-me/CLIProxyAPI/v6/internal/usage"
+)
+
+func oauthSettingsService(h *Handler) *oauthsettings.Service {
+	if h == nil {
+		return oauthsettings.NewService(nil)
+	}
+	return oauthsettings.NewService(h.cfg)
+}
+
+// oauth-excluded-models: map[string][]string
+func (h *Handler) GetOAuthExcludedModels(c *gin.Context) {
+	c.JSON(200, gin.H{"oauth-excluded-models": oauthSettingsService(h).ExcludedModels()})
+}
+
+func (h *Handler) PutOAuthExcludedModels(c *gin.Context) {
+	data, err := c.GetRawData()
+	if err != nil {
+		c.JSON(400, gin.H{"error": "failed to read body"})
+		return
+	}
+	var entries map[string][]string
+	if err = json.Unmarshal(data, &entries); err != nil {
+		var wrapper struct {
+			Items map[string][]string `json:"items"`
+		}
+		if err2 := json.Unmarshal(data, &wrapper); err2 != nil {
+			c.JSON(400, gin.H{"error": "invalid body"})
+			return
+		}
+		entries = wrapper.Items
+	}
+	setting := oauthSettingsService(h).SetExcludedModels(entries)
+	h.persistRuntimeSetting(c, usage.RuntimeSettingOAuthExcludedModels, setting)
+}
+
+func (h *Handler) PatchOAuthExcludedModels(c *gin.Context) {
+	var body struct {
+		Provider *string  `json:"provider"`
+		Models   []string `json:"models"`
+	}
+	if err := c.ShouldBindJSON(&body); err != nil || body.Provider == nil {
+		c.JSON(400, gin.H{"error": "invalid body"})
+		return
+	}
+	setting, err := oauthSettingsService(h).PatchExcludedModels(*body.Provider, body.Models)
+	if err == oauthsettings.ErrInvalidProvider {
+		c.JSON(400, gin.H{"error": "invalid provider"})
+		return
+	}
+	if err == oauthsettings.ErrProviderNotFound {
+		c.JSON(404, gin.H{"error": "provider not found"})
+		return
+	}
+	h.persistRuntimeSetting(c, usage.RuntimeSettingOAuthExcludedModels, setting)
+}
+
+func (h *Handler) DeleteOAuthExcludedModels(c *gin.Context) {
+	setting, err := oauthSettingsService(h).DeleteExcludedModels(c.Query("provider"))
+	if err == oauthsettings.ErrInvalidProvider {
+		c.JSON(400, gin.H{"error": "missing provider"})
+		return
+	}
+	if err == oauthsettings.ErrProviderNotFound {
+		c.JSON(404, gin.H{"error": "provider not found"})
+		return
+	}
+	h.persistRuntimeSetting(c, usage.RuntimeSettingOAuthExcludedModels, setting)
+}
+
+// oauth-model-alias: map[string][]OAuthModelAlias
+func (h *Handler) GetOAuthModelAlias(c *gin.Context) {
+	c.JSON(200, gin.H{"oauth-model-alias": oauthSettingsService(h).ModelAlias()})
+}
+
+func (h *Handler) PutOAuthModelAlias(c *gin.Context) {
+	data, err := c.GetRawData()
+	if err != nil {
+		c.JSON(400, gin.H{"error": "failed to read body"})
+		return
+	}
+	var entries map[string][]config.OAuthModelAlias
+	if err = json.Unmarshal(data, &entries); err != nil {
+		var wrapper struct {
+			Items map[string][]config.OAuthModelAlias `json:"items"`
+		}
+		if err2 := json.Unmarshal(data, &wrapper); err2 != nil {
+			c.JSON(400, gin.H{"error": "invalid body"})
+			return
+		}
+		entries = wrapper.Items
+	}
+	setting := oauthSettingsService(h).SetModelAlias(entries)
+	h.persistRuntimeSetting(c, usage.RuntimeSettingOAuthModelAlias, setting)
+}
+
+func (h *Handler) PatchOAuthModelAlias(c *gin.Context) {
+	var body struct {
+		Provider *string                  `json:"provider"`
+		Channel  *string                  `json:"channel"`
+		Aliases  []config.OAuthModelAlias `json:"aliases"`
+	}
+	if errBindJSON := c.ShouldBindJSON(&body); errBindJSON != nil {
+		c.JSON(400, gin.H{"error": "invalid body"})
+		return
+	}
+	channelRaw := ""
+	if body.Channel != nil {
+		channelRaw = *body.Channel
+	} else if body.Provider != nil {
+		channelRaw = *body.Provider
+	}
+	setting, err := oauthSettingsService(h).PatchModelAlias(channelRaw, body.Aliases)
+	if err == oauthsettings.ErrInvalidChannel {
+		c.JSON(400, gin.H{"error": "invalid channel"})
+		return
+	}
+	if err == oauthsettings.ErrChannelNotFound {
+		c.JSON(404, gin.H{"error": "channel not found"})
+		return
+	}
+	h.persistRuntimeSetting(c, usage.RuntimeSettingOAuthModelAlias, setting)
+}
+
+func (h *Handler) DeleteOAuthModelAlias(c *gin.Context) {
+	channel := strings.ToLower(strings.TrimSpace(c.Query("channel")))
+	if channel == "" {
+		channel = strings.ToLower(strings.TrimSpace(c.Query("provider")))
+	}
+	if channel == "" {
+		c.JSON(400, gin.H{"error": "missing channel"})
+		return
+	}
+	setting, err := oauthSettingsService(h).DeleteModelAlias(channel)
+	if err == oauthsettings.ErrChannelNotFound {
+		c.JSON(404, gin.H{"error": "channel not found"})
+		return
+	}
+	h.persistRuntimeSetting(c, usage.RuntimeSettingOAuthModelAlias, setting)
+}
